+++ Common/src/main/java/de/markusbordihn/easymobfarm/loot/LootManager.java
@@ -280,18 +280,20 @@ public class LootManager {
     float additionalLuck = 0;
     int additionalRolls = 0;
     for (EnhancementItem enhancement : enhancements) {
-      if (enhancement instanceof SwordEnhancementItem && MobFarmConfig.enableSwordEnhancement) {
+      if (enhancement instanceof SwordEnhancementItem
+          && MobFarmConfig.isSwordEnhancementEnabled()) {
         setSwordEnhancementParameters(lootContextBuilder, fakePlayer, serverLevel);
         additionalLuck += MobFarmConfig.swordEnhancementAdditionalLuck;
       }
-      if (enhancement instanceof KnifeEnhancementItem && MobFarmConfig.enableKnifeEnhancement) {
+      if (enhancement instanceof KnifeEnhancementItem
+          && MobFarmConfig.isKnifeEnhancementEnabled()) {
         setKnifeEnhancementParameters(lootContextBuilder, fakePlayer, serverLevel);
         additionalLuck += MobFarmConfig.knifeEnhancementAdditionalLuck;
       }
-      if (enhancement instanceof LootEnhancementItem && MobFarmConfig.enableLootEnhancement) {
+      if (enhancement instanceof LootEnhancementItem && MobFarmConfig.isLootEnhancementEnabled()) {
         additionalRolls += MobFarmConfig.lootEnhancementAdditionalRolls;
       }
-      if (enhancement instanceof LuckEnhancementItem && MobFarmConfig.enableLuckEnhancement) {
+      if (enhancement instanceof LuckEnhancementItem && MobFarmConfig.isLuckEnhancementEnabled()) {
         additionalLuck += MobFarmConfig.luckEnhancementAdditionalLuck;
       }
     }
@@ -414,8 +416,11 @@ public class LootManager {
       }
       if (entityType == EntityType.BEE
           && drop.is(Items.HONEYCOMB)
-          && MobFarmConfig.enableHoneyExtractorEnhancement
-          && enhancements.stream().anyMatch(e -> e instanceof HoneyExtractorEnhancementItem)) {
+          && enhancements.stream()
+              .anyMatch(
+                  enhancement ->
+                      enhancement instanceof HoneyExtractorEnhancementItem
+                          && MobFarmConfig.isEnhancementEnabled(enhancement))) {
         drops.add(new ItemStack(Items.HONEY_BOTTLE));
       } else {
         drops.add(drop.copy());
@@ -455,7 +460,7 @@ public class LootManager {
     ResourceLocation lootTableLocation = livingEntity.getType().getDefaultLootTable();
     for (EnhancementItem enhancement : enhancements) {
       if (enhancement instanceof SheepEnhancementItem
-          && MobFarmConfig.enableSheepEnhancement
+          && MobFarmConfig.isEnhancementEnabled(enhancement)
           && livingEntity instanceof Sheep sheep) {
         DyeColor color = sheep.getColor();
         lootTableLocation = new ResourceLocation("minecraft", "entities/sheep/" + color.getName());
@@ -556,7 +561,7 @@ public class LootManager {
 
       // Handle Experience enhancement
       if (enhancement instanceof ExperienceEnhancementItem experienceEnhancementItem
-          && MobFarmConfig.enableExperienceEnhancement
+          && MobFarmConfig.isEnhancementEnabled(enhancement)
           && random.nextInt(experienceEnhancementItem.experienceDropChance()) == 0
           && ExperienceManager.shouldDropExperience(livingEntity)) {
         int experience = ExperienceManager.getExperienceReward(livingEntity);
@@ -574,15 +579,15 @@ public class LootManager {
       // Handle Mob specific enhancements
       if (livingEntity instanceof Bee) {
         if (enhancement instanceof HoneyHarvesterFrameEnhancementItem
-            && MobFarmConfig.enableHoneyHarvesterFrameEnhancement
+            && MobFarmConfig.isEnhancementEnabled(enhancement)
             && random.nextInt(4) == 0) {
           drops.add(new ItemStack(Items.HONEYCOMB));
         } else if (enhancement instanceof HoneyExtractorEnhancementItem
-            && MobFarmConfig.enableHoneyExtractorEnhancement
+            && MobFarmConfig.isEnhancementEnabled(enhancement)
             && random.nextInt(4) == 0) {
           drops.add(new ItemStack(Items.HONEY_BOTTLE));
         } else if (enhancement instanceof PollenTrapEnhancementItem
-            && MobFarmConfig.enablePollenTrapEnhancement
+            && MobFarmConfig.isEnhancementEnabled(enhancement)
             && random.nextInt(5) == 0) {
           if (random.nextFloat() < 0.3f) {
             drops.add(getRandomFlower());
@@ -593,7 +598,7 @@ public class LootManager {
       } else if (livingEntity instanceof Cow) {
         // Handle MilkExtractor enhancement (50% chance)
         if (enhancement instanceof MilkExtractorEnhancementItem
-            && MobFarmConfig.enableMilkExtractorEnhancement
+            && MobFarmConfig.isEnhancementEnabled(enhancement)
             && random.nextInt(2) == 0) {
           Item milkBottle;
           if (CompatConstants.MOD_FARMERS_DELIGHT_LOADED) {
@@ -614,7 +619,7 @@ public class LootManager {
       // Handle egg drops for chicken entities with Egg Collector enhancement (50% chance)
       if (livingEntity instanceof Chicken) {
         if (enhancement instanceof EggCollectorEnhancementItem
-            && MobFarmConfig.enableEggCollectorEnhancement
+            && MobFarmConfig.isEnhancementEnabled(enhancement)
             && random.nextInt(2) == 0) {
           drops.add(new ItemStack(Items.EGG));
         }
@@ -645,6 +650,7 @@ public class LootManager {
 
         // Adding additional Frog Light drop with a 50% chance with frog catalyst enhancement.
         if (enhancement instanceof FrogCatalystEnhancementItem frogCatalystEnhancementItem
+            && MobFarmConfig.isEnhancementEnabled(enhancement)
             && random.nextInt(2) == 0) {
           FrogCatalystType frogCatalystType = frogCatalystEnhancementItem.getFrogCatalystType();
           // Handle basic frog catalyst drops based on type.
@@ -685,7 +691,8 @@ public class LootManager {
       NonNullList<ItemStack> drops) {
     boolean hasKnifeEnhancement = false;
     for (EnhancementItem enhancement : enhancements) {
-      if (enhancement instanceof KnifeEnhancementItem) {
+      if (enhancement instanceof KnifeEnhancementItem
+          && MobFarmConfig.isEnhancementEnabled(enhancement)) {
         hasKnifeEnhancement = true;
         break;
       }