+++ Common/src/main/java/de/markusbordihn/easymobfarm/menu/slots/EnhancementSlot.java
@@ -19,6 +19,8 @@
 
 package de.markusbordihn.easymobfarm.menu.slots;
 
+import de.markusbordihn.easymobfarm.block.entity.MobFarmBlockEntity;
+import de.markusbordihn.easymobfarm.config.MobFarmConfig;
 import de.markusbordihn.easymobfarm.item.upgrade.EnhancementItem;
 import de.markusbordihn.easymobfarm.menu.MobFarmSlot;
 import de.markusbordihn.easymobfarm.network.components.TextComponent;
@@ -29,19 +31,25 @@ import net.minecraft.world.item.ItemStack;
 
 public class EnhancementSlot extends MobFarmSlot {
 
+  private final int containerSlotIndex;
+
   public EnhancementSlot(Container container, int index, int x, int y) {
     super(container, index, x, y);
+    this.containerSlotIndex = index;
   }
 
   @Override
   public boolean mayPlace(ItemStack itemStack) {
-    if (itemStack.isEmpty() || hasItem()) {
+    if (itemStack.isEmpty()) {
       return false;
     }
 
     Item item = itemStack.getItem();
-    if (item instanceof EnhancementItem) {
-      return true;
+    if (item instanceof EnhancementItem enhancementItem) {
+      if (container instanceof MobFarmBlockEntity mobFarmBlockEntity) {
+        return mobFarmBlockEntity.canAddEnhancementItem(enhancementItem, this.containerSlotIndex);
+      }
+      return MobFarmConfig.isEnhancementEnabled(enhancementItem);
     }
 
     log.debug(